fixing rotation issue
